Add files via upload
This is my Project Titled "DRIVER DROWSINESS DETECTION AND ALERTING SYSTEM USING ARDUINO"
This projects has been developed to reduce accidents that occur due to the drowsiness of the driver while driving.This projects works using EYE BLINK SENSOR which captures the eye movement of the driver while driving and gives alert (BEEP SOUND) to the driver to make him alert. This system is implemented using the ARDUINO in which I have implemented the code to generate.